Remove Tile Dictionary
Going to move toward using a grid system to create the map in the code. This may make it harder to work with in unity, but I don't like the idea of determining the data of the map from the presented layout, I would rather have it work the other way around.